+++ ouroboros-consensus-cardano/tools/db-analyser/HasAnalysis.hs
@@ -8,7 +8,7 @@ import           Options.Applicative
 
 import           Ouroboros.Consensus.Block
 import           Ouroboros.Consensus.Node.ProtocolInfo
-import           Ouroboros.Consensus.Storage.ChainDB.Serialisation (SizeInBytes)
+import           Ouroboros.Consensus.Storage.Serialisation (SizeInBytes)
 
 {-------------------------------------------------------------------------------
   HasAnalysis
