+++ ouroboros-consensus-shelley-test/src/Test/Consensus/Shelley/Examples.hs
@@ -57,7 +57,7 @@ import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Ledger.Extended
 import           Ouroboros.Consensus.Ledger.SupportsMempool
 import           Ouroboros.Consensus.Protocol.Abstract
-import           Ouroboros.Consensus.Storage.ChainDB.Serialisation
+import           Ouroboros.Consensus.Storage.Serialisation
 import           Ouroboros.Consensus.Util.Time
 
 import           Cardano.Ledger.Crypto (ADDRHASH, DSIGN, VRF)
