+++ ouroboros-consensus-test/test-storage/Test/Ouroboros/Storage/ImmutableDB/Mock.hs
@@ -10,11 +10,11 @@ import           Ouroboros.Consensus.Block
 import           Ouroboros.Consensus.Util ((...:), (.:))
 import           Ouroboros.Consensus.Util.IOLike
 
-import           Ouroboros.Consensus.Storage.ChainDB.Serialisation
 import           Ouroboros.Consensus.Storage.Common (BlockComponent)
 import           Ouroboros.Consensus.Storage.ImmutableDB.API
 import           Ouroboros.Consensus.Storage.ImmutableDB.Chunks
 import           Ouroboros.Consensus.Storage.ImmutableDB.Error
+import           Ouroboros.Consensus.Storage.Serialisation
 
 import           Test.Ouroboros.Storage.ImmutableDB.Model
