+++ ouroboros-consensus/src/Ouroboros/Consensus/HardFork/Combinator/Serialisation/SerialiseDisk.hs
@@ -19,7 +19,7 @@ import           Ouroboros.Consensus.TypeFamilyWrappers
 import           Ouroboros.Consensus.Util ((.:))
 
 import           Ouroboros.Consensus.Storage.ChainDB
-import           Ouroboros.Consensus.Storage.ChainDB.Serialisation
+import           Ouroboros.Consensus.Storage.Serialisation
 
 instance SerialiseHFC xs => SerialiseDiskConstraints  (HardForkBlock xs)
 instance SerialiseHFC xs => LgrDbSerialiseConstraints (HardForkBlock xs)
